requirements.txt
# Production requirements
flask==3.1.0
flask-sqlalchemy==3.1.1
flask-login==0.6.3
flask-mail==0.10.0
flask-wtf==1.2.2
flask-compress==1.17
gunicorn==23.0.0
psycopg2-binary==2.9.10
sqlalchemy==2.0.40
werkzeug==3.1.3

# Email & Communication
sendgrid==6.11.0
email-validator==2.2.0

# AI & Chatbot
openai==1.78.1

# Payment Processing
stripe==12.0.0

# Utilities
python-slugify==8.0.4
python-dotenv==1.0.0

# Development dependencies
pytest==7.4.0
pytest-flask==1.2.0
pytest-mock==3.11.1
black==23.7.0
flake8==6.0.0
